Remove the original sort, substitute with external sort
